Phase 3: real STT/TTS with pipelined LLM->TTS for the verdict
Booth now does the full audio loop end-to-end against the live Spark:
mic plea -> Parakeet STT -> Qwen3.6 streaming verdict -> Kokoro TTS
streamed sentence-by-sentence to the browser as PCM. Verified guilty
intensity 4 with FIRE 200 firing after audio drains.

- inference/client.rs: transcribe() multipart upload to
  /v1/audio/transcriptions; synth_pcm_stream() for /v1/audio/speech
  with response_format=pcm.
- inference/stt.rs: real STT path; empty/failed transcript falls back
  to "[no defense offered]".
- inference/tts.rs: real TTS for the charge (single-utterance streaming)
  with strip_markers() helper that drops VERDICT:/INTENSITY: lines so
  they're never spoken. Headless self-ack estimates playback duration
  from PCM byte count so the state machine doesn't wedge without a
  browser.
- inference/verdict.rs: pipelined LLM -> sentence buffer -> ordered
  TTS worker per architecture §6.3. Sentences shorter than 25 chars
  glue onto the next one to avoid clipped Kokoro output. Stops feeding
  TTS once VERDICT: appears mid-stream. Sets pre_announced=true on the
  resulting Verdict so the state machine skips a redundant Speak.
- state_machine: Verdict gains pre_announced field; begin_pronouncing
  drops the Speak command when set. Transcribe takes Vec<u8> opaque
  bytes (Parakeet handles standard formats; no client-side resampling).
- display: broadcast channel now carries DisplayMessage::{Json,Binary};
  ws_handler forwards binary frames; binary uploads from the client
  accumulate in a per-session plea_buffer flushed on
  plea_audio_complete.
- display/events.rs: new TtsEnd marker so the frontend knows when to
  schedule tts_finished after the queued audio drains.
- frontend/audio.ts: AudioContext-based gapless PCM playback queue
  (Int16 -> Float32 -> AudioBuffer -> BufferSource scheduled at the
  tail). MediaRecorder mic capture (webm/opus) with single-blob upload.
  Honors browser autoplay rules via resumeAudio() in the Start handler.
- frontend/ws.ts: routes binary frames to enqueuePcmFrame when the
  preceding JSON event was tts_audio; drives mic via
  start_plea_recording / stop_plea_recording display events.
- frontend/App.tsx: streaming deliberation pane (italic Georgia)
  renders deliberation_token text live as the judge speaks.

+++ orchestrator/src/state_machine/events.rs
@@ -8,7 +8,7 @@ pub enum Event {
     ChargeReady(String),
     #[allow(dead_code)]
     ChargeFailed(String),
-    PleaAudioReceived(Vec<f32>),
+    PleaAudioReceived(Vec<u8>),
     PleaTimeout,
     TranscriptReady(String),
     #[allow(dead_code)]
